day 80: refactor same problem from day 12 extra problem
