+++ day80/problem.go
@@ -1,17 +1,9 @@
-package deepest
+package day80
 
 // BinaryTree represents a tree with a left and right subtree.
 type BinaryTree struct {
 	// val         int // value doesn't matter here.
 	left, right *BinaryTree
-}
-
-// CountNodes tallies the total number of nodes in the binary tree.
-func CountNodes(root *BinaryTree) int {
-	if root == nil {
-		return 0
-	}
-	return CountNodes(root.left) + CountNodes(root.right) + 1
 }
 
 // Deepest returns the deepest node in the binary tree.
